Enter payment to UC order during payment confirmation
Entering payments to the Ubercart order when a user returns to the
success page turned out to be problematic, because some users will
choose not to return to the site after they see a message on the EPay
page that states that the payment was successful.

Fortunately, we are able to enter payment to an order when we process a
payment confirmation from the EPay server.